tests incluing emailAddresses added for filter alerts
